requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask_Migrate==2.7.0
Flask_SQLAlchemy==2.5.1
psycopg2==2.9.5
psycopg2_binary==2.9.5
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
